+++ ui/index.js
@@ -6,9 +6,10 @@ const configureStore = require('./app/store')
 const txHelper = require('./lib/tx-helper')
 module.exports = launchApp
 
+let debugMode = window.METAMASK_DEBUG
 const log = require('loglevel')
 window.log = log
-log.setLevel('warn')
+log.setLevel(debugMode ? 'debug' : 'warn')
 
 function launchApp (opts) {
   var accountManager = opts.accountManager
@@ -36,7 +37,7 @@ function startApp (metamaskState, accountManager, opts) {
   })
 
   // if unconfirmed txs, start on txConf page
-  var unapprovedTxsAll = txHelper(metamaskState.unapprovedTxs, metamaskState.unapprovedMsgs, metamaskState.network)
+  var unapprovedTxsAll = txHelper(metamaskState.unapprovedTxs, metamaskState.unapprovedMsgs, metamaskState.unapprovedPersonalMsgs, metamaskState.network)
   if (unapprovedTxsAll.length > 0) {
     store.dispatch(actions.showConfTxPage())
   }