+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 ## Current Master
 
 - Add two most recently used custom RPCs to network dropdown menu.
+- Add personal_sign method support.
+- Add ability to customize gas and gasPrice on the transaction approval screen.
 
 ## 3.3.0 2017-2-20
 
-- Add personal_sign and personal_ecRecover support.
 - net_version has been made synchronous.
 - Test suite for migrations expanded.
 - Network now changeable from lock screen.